+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/constant/MetadataConstant.java
@@ -17,8 +17,6 @@
 
 package com.tencent.cloud.common.constant;
 
-import org.springframework.core.Ordered;
-
 /**
  * Constant for metadata.
  *
@@ -46,27 +44,6 @@ public final class MetadataConstant {
 
 	private MetadataConstant() {
 
-	}
-
-	/**
-	 * Order of filter, interceptor, ...
-	 */
-	public static class OrderConstant {
-
-		/**
-		 * Order of filter.
-		 */
-		public static final int WEB_FILTER_ORDER = Ordered.HIGHEST_PRECEDENCE + 9;
-
-		/**
-		 * Order of MetadataFirstFeignInterceptor.
-		 */
-		public static int METADATA_FIRST_FEIGN_INTERCEPTOR_ORDER = Ordered.HIGHEST_PRECEDENCE + 1;
-
-		/**
-		 * Order of Metadata2HeaderInterceptor.
-		 */
-		public static int METADATA_2_HEADER_INTERCEPTOR_ORDER = Ordered.LOWEST_PRECEDENCE;
 	}
 
 	/**