+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - refactor:refactor Polaris registration.
 - feat:added automatic optimization for dynamic config refresh type.
 - fix:fix gray release examples bug.
+- fix:fix router label feign interceptor order.